DAS-2064: rename/relocate mapping.py -> xarray.core.datatree_mapping.py
DAS-2064: fix circular import issue.
