+++ xarray/datatree_/datatree/ops.py
@@ -2,7 +2,7 @@ import textwrap
 
 from xarray import Dataset
 
-from .mapping import map_over_subtree
+from xarray.core.datatree_mapping import map_over_subtree
 
 """
 Module which specifies the subset of xarray.Dataset's API which we wish to copy onto DataTree.
